Move VoidToMonostate to its own header
This is so it can be used elsewhere.
